Fix rebase issues and split into library and cli/winit binaries
