+++ piet-gpu/Cargo.toml
@@ -6,6 +6,14 @@ description = "A compute-centric GPU 2D renderer."
 license = "MIT/Apache-2.0"
 edition = "2018"
 
+[[bin]]
+name = "cli"
+path = "bin/cli.rs"
+
+[[bin]]
+name = "winit"
+path = "bin/winit.rs"
+
 [dependencies.piet-gpu-hal]
 path = "../piet-gpu-hal"
 
@@ -18,4 +26,4 @@ piet = "0.0.12"
 png = "0.16.2"
 rand = "0.7.3"
 roxmltree = "0.11"
-winit = "0.22"
+winit = "0.22"
