+++ piet-gpu/shader/kernel4.comp
@@ -24,7 +24,7 @@ layout(set = 0, binding = 2) buffer FillSegBuf {
     uint[] fill_seg;
 };
 
-layout(rgba8, set = 0, binding = 2) uniform writeonly image2D image;
+layout(rgba8, set = 0, binding = 3) uniform writeonly image2D image;
 
 #include "ptcl.h"
 #include "segment.h"
